requirements.txt
langid==1.1.6
kenlm==0.0.0
torch==0.1.12.post2
tabulate==0.7.7
more_itertools==3.2.0
six==1.10.0
numpy==1.13.0
matplotlib==2.0.0
tqdm==4.11.2
tensorboard_logger==0.0.3
nltk==3.2.2
scipy==0.19.1
python_rex==0.4
lightgbm==2.0.6
num2words==0.5.5
forked_path==0.2.3
rex==0.9.0
scikit_learn==0.19.0
xgboost==0.6
